pyproject.toml
[project]
name = "fastapi-radar"
version = "0.3.4"
description = "A debugging dashboard for FastAPI applications with real-time monitoring"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{ name = "Arif Dogan", email = "me@arif.sh" }]
keywords = [
    "fastapi",
    "debugging",
    "monitoring",
    "dashboard",
    "development-tools",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastapi",
    "pydantic",
    "starlette",
    "tortoise-orm>=0.20.0",
    "aiosqlite>=0.17.0",
]

[project.urls]
"Homepage" = "https://github.com/doganarif/fastapi-radar"
"Bug Reports" = "https://github.com/doganarif/fastapi-radar/issues"
"Source" = "https://github.com/doganarif/fastapi-radar"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.1",
    "uvicorn[standard]>=0.23.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "httpx>=0.24.0",
    "bandit[toml]>=1.7.5",
    "safety>=2.3.5",
    "pre-commit>=3.3.3",
]

release = [
    "build>=0.10.0",
    "twine>=4.0.2"
]


[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["fastapi_radar*"]

[tool.setuptools.package-data]
fastapi_radar = ["dashboard/dist/**/*"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | dashboard/node_modules
  | dashboard/dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = [
    "*/dashboard/node_modules/*",
    "*/dashboard/dist/*",
]

# Bandit security configuration
[tool.bandit]
exclude_dirs = [
    "tests",
    "dashboard/node_modules",
    "dashboard/dist",
]
skips = ["B101"]

[tool.bandit.assert_used]
skips = ["**/test_*.py", "**/conftest.py"]
